+++ llvm/lib/MC/MCObjectFileInfo.cpp
@@ -57,8 +57,8 @@ void MCObjectFileInfo::initMachOMCObjectFileInfo(const Triple &T) {
 
   EHFrameSection = Ctx->getMachOSection(
       "__TEXT", "__eh_frame",
-      MachO::S_COALESCED | MachO::S_ATTR_NO_TOC |
-          MachO::S_ATTR_STRIP_STATIC_SYMS | MachO::S_ATTR_LIVE_SUPPORT,
+      static_cast<unsigned int>(MachO::S_COALESCED) | static_cast<unsigned int>(MachO::S_ATTR_NO_TOC) |
+          static_cast<unsigned int>(MachO::S_ATTR_STRIP_STATIC_SYMS) | static_cast<unsigned int>(MachO::S_ATTR_LIVE_SUPPORT),
       SectionKind::getReadOnly());
 
   if (T.isOSDarwin() &&
@@ -153,8 +153,8 @@ void MCObjectFileInfo::initMachOMCObjectFileInfo(const Triple &T) {
   if (ArchTy == Triple::ppc || ArchTy == Triple::ppc64) {
     TextCoalSection
       = Ctx->getMachOSection("__TEXT", "__textcoal_nt",
-                             MachO::S_COALESCED |
-                             MachO::S_ATTR_PURE_INSTRUCTIONS,
+                             static_cast<unsigned int>(MachO::S_COALESCED) |
+                             static_cast<unsigned int>(MachO::S_ATTR_PURE_INSTRUCTIONS),
                              SectionKind::getText());
     ConstTextCoalSection
       = Ctx->getMachOSection("__TEXT", "__const_coal",
